+++ common/scripts/definitions/maps/normal.ts
@@ -137,7 +137,8 @@ export const NormalMap:MapDef={
         ],
         //Throwables
         "throwables":[
-            {item:"frag_grenade",count:1,weight:100},
+            {item:"frag_grenade",count:1,weight:50},
+            {item:"smoke_grenade",count:2,weight:40},
             {item:"mirv_grenade",count:1,weight:20},
             {item:"blue_flare",count:1,weight:1},
             {item:"red_flare",count:1,weight:1},